ransomware_vault-main/requirements.txt
playsound==1.3.0
Pillow==7.0.0
PyPDF2==1.26.0
python-docx==0.8.11
python-magic==0.4.24
pytest==6.2.5
